+++ finitevolx/_src/mask/__init__.py
@@ -6,6 +6,13 @@ from finitevolx._src.mask.base import (
     StencilCapability3D,
 )
 from finitevolx._src.mask.cartesian import Mask1D, Mask2D, Mask3D
+from finitevolx._src.mask.utils import (
+    count_contiguous,
+    dilate_mask,
+    h_from_pooled,
+    make_sponge,
+    pool_bool,
+)
 
 __all__ = [
     # Stencil capability
@@ -16,4 +23,10 @@ __all__ = [
     "Mask1D",
     "Mask2D",
     "Mask3D",
+    # Mask construction primitives
+    "count_contiguous",
+    "dilate_mask",
+    "h_from_pooled",
+    "make_sponge",
+    "pool_bool",
 ]